+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/forging/ForgeHammerItem.java
@@ -40,7 +40,7 @@ public class ForgeHammerItem extends TieredItem implements Vanishable {
         super(tier, pProperties);
         this.useCooldown = useCooldown;
         ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
-        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", 10 + getTier().getAttackDamageBonus(), AttributeModifier.Operation.ADDITION));
+        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", 3.5 + getTier().getAttackDamageBonus(), AttributeModifier.Operation.ADDITION));
         builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Tool modifier", -(4d - 0.35), AttributeModifier.Operation.ADDITION));
         builder.put(ForgeMod.ENTITY_REACH.get(), new AttributeModifier(ENTITY_REACH_UUID, "Tool modifier", -1d, AttributeModifier.Operation.ADDITION));
         this.defaultModifiers = builder.build();