+++ changelog.md
@@ -7,6 +7,7 @@
 * Increased Gravity Defying and M.A. Jump cost
 * Soul Steel now requires Netherite Scrap to be made
   * Can also be made in a crafting table like Netherite
+  * With this, Netherite Scrap has been renamed to Ancient Scrap
 * Like the base mod, materials now have a percentage increase/decrease in attack speed
 * Fixed missing Quaron Nugget recipe
 * Fixed enchanting table screen having no background